+++ temba/flows/models.py
@@ -1114,6 +1114,7 @@ class FlowSession(models.Model):
     wait_started_on = models.DateTimeField(null=True)  # when it started waiting
     timeout_on = models.DateTimeField(null=True)  # when it should timeout (set by courier when last msg is sent)
     wait_expires_on = models.DateTimeField(null=True)  # when waiting run can be expired
+    wait_resume_on_expire = models.BooleanField(null=True)  # whether wait expiration can resume a parent run
 
     # the flow of the waiting run
     current_flow = models.ForeignKey("flows.Flow", related_name="sessions", null=True, on_delete=models.PROTECT)
@@ -1263,9 +1264,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
     # TODO to be replaced by FlowSession.wait_expires_on
     expires_on = models.DateTimeField(null=True)
 
-    # TODO drop once mailroom is no longer writing
-    events = JSONField(null=True)
-
     def release(self, delete_reason=None):
         """
         Permanently deletes this flow run
@@ -1274,9 +1272,6 @@ class FlowRun(RequireUpdateFieldsMixin, models.Model):
             if delete_reason:
                 self.delete_reason = delete_reason
                 self.save(update_fields=["delete_reason"])
-
-            # and any recent runs
-            self.recent_runs.all().delete()
 
             if (
                 delete_reason == FlowRun.DELETE_FOR_USER
@@ -1585,23 +1580,6 @@ class FlowPathCount(SquashableModel):
 
     class Meta:
         index_together = ["flow", "from_uuid", "to_uuid", "period"]
-
-
-class FlowPathRecentRun(models.Model):
-    """
-    TODO drop once we stop writing these
-    """
-
-    id = models.BigAutoField(primary_key=True)
-    from_uuid = models.UUIDField()
-    from_step_uuid = models.UUIDField()
-    to_uuid = models.UUIDField()
-    to_step_uuid = models.UUIDField()
-    run = models.ForeignKey(FlowRun, on_delete=models.PROTECT, related_name="recent_runs")
-    visited_on = models.DateTimeField(default=timezone.now)
-
-    class Meta:
-        indexes = [models.Index(fields=["from_uuid", "to_uuid", "-visited_on"])]
 
 
 class FlowNodeCount(SquashableModel):