+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+v7.1.40
+----------
+ * Drop FlowRun.events and FlowPathRecentRun
+
+v7.1.39
+----------
+ * Include qrious.js script
+ * Add FlowSession.wait_resume_on_expire
+ * Add Msg.flow
+
 v7.1.38
 ----------
  * Replace uses of deprecated Django functions